Move non target_quality specific functions to vmaf
